Add `ValidatorActor` to future handler (#140)
Renames `Validator` to `ValidatorActor`. Slight refactor to the `ractor::Actor` impl for `ValidatorActor` to use the `ValidatorCore` that's instantiated in the `main` function, this will also reuse the same data instead of creating a new default `ValidatorCore` on `pre_start`. Implements `ActorExt` for `ValidatorActor` and adds a test to ensure its effectiveness. Updates `ValidatorError` to make error conversion easier in the abstracted `ValidatorActor` associated functions.