Remove unnecessary async await in EoServerActor (#143)
While refactoring the `EoServer` to be added to the future handler, I realized the `async` functions that are called are unnecessary. This refactor just renames `EoServer` to `EoServerActor` and removes the unused async await points from the `ractor::Actor::handle` method. It also refactors methods that take `self` but do not use `self` into associated functions.