Add more robust error handling to `Actor` methods (#150)
This does as much as I think is doable right now for the `BatcherActor`, and actually completely resolves the need for end point logging in the `DaClientActor`. This adds a lot of setup code for future methods, and gives us some pretty powerful abstractions not only for acquiring an `ActorRef` from the registry but also for logging solid error types (not just in place logging!), which can almost entirely replace `.map_err` while retaining the intention of the writer.