Add testing framework for events that update `Account` (#210)
Closes #205 

Adds the tests necessary to prove events that update `Account` function
as intended. Also adds the ability to test with in-memory key-value
storage (`HashMap`) which relieves the overhead of needing a
`tikv_client` running. The `MinimalNode` testing type can eventually be
moved to its own module and expanded upon if needed, which makes testing
node processes outside of RPC calls, IPFS and other compute related
stuff (outside of the scope of these tests) relatively easy.